Module type Mirage_crypto_rng_mirage.S

type g = Mirage_crypto_rng.g

A generator (PRNG) with its state.

module Entropy : sig ... end

Entropy sources and collection

val generate_into : ?g:g -> bytes -> ?off:int -> int -> unit

generate_into ~g buf ~off len invokes generate_into on g or default generator. The random data is put into buf starting at off (defaults to 0) with len bytes.

val generate : ?g:g -> int -> string

Invoke generate_into on g or default generator and a freshly allocated string.

val accumulate : g option -> Entropy.source -> [ `Acc of string -> unit ]

accumulate g source is a function data -> unit to feed entropy to the RNG. This is useful if your system has a special entropy source.